refactor: create Projects section
