+++ src/pages/landing-page/sections/projects/projects-mock.ts
